Address review items: docs fixes, sweep API cleanup, module splits
- README.md: add T-Rex README link, include README.md in trex/ structure
  diagram, correct brachiosaurus action dim (30 -> 26) to match the env.
- website/docs/training/hyperparameters.md: add a Tuning Playbook with
  per-stage symptom -> remedy tables and a minimal iteration workflow.
- sweep/__init__.py: drop 21 underscore-prefixed test-only helpers from
  __all__. Names remain importable for the test suite, but the advertised
  public API is now the non-underscore surface only.
- train_base.py: extract GCS TensorBoard buffering helpers (_is_gcs_path,
  _make_local_tb_dir, _sync_tb_to_gcs) to new tb_sync.py; re-export for
  backward compatibility.
- jax_trainer.py: extract TrainerState, TrainingHook, StopTraining to
  new jax_trainer_types.py; re-export for backward compatibility.
