+++ environments/shared/scripts/sweep/__init__.py
@@ -74,6 +74,10 @@ from .state import (
 from .submit import _is_retryable_gcp_error, _normalize_accelerator_type, _submit_stage_sweep, _validate_machine_type
 from .trial import _hpt_arg_to_override, _parse_hpt_extra_args, run_trial
 
+# Public API. Underscore-prefixed helpers (``_load_sweep_state``,
+# ``_submit_stage_sweep``, etc.) are intentionally imported above so existing
+# test modules keep working but are NOT re-exported here — consumers should
+# import them from their defining submodule if genuinely needed.
 __all__ = [
     "DriveProgressLogCallback",
     "ExperimentStateSyncCallback",
@@ -81,41 +85,10 @@ __all__ = [
     "RayTuneReportCallback",
     "SweepStageError",
     "TrialTerminationCallback",
-    "_SweepJobFailed",
-    "_best_trial_model_path",
-    "_best_trial_model_path_any",
-    "_collect_trial_results",
-    "_eager_refresh",
-    "_evaluate_curriculum_gate",
-    "_extract_thresholds",
     "apply_sampled_config",
     "build_search_space",
     "collect_ray_results",
     "collect_results_from_disk",
-    "_hpt_arg_to_override",
-    "_parse_hpt_extra_args",
-    "_is_per_stage",
-    "_is_retryable_gcp_error",
-    "_load_sweep_state",
-    "_normalize_accelerator_type",
-    "_resolve_search_space",
-    "_save_sweep_state",
-    "_search_space_for_stage",
-    "_settings_for_stage",
-    "_split_stage_block",
-    "_submit_stage_sweep",
-    "_sweep_state_local_path",
-    "_validate_machine_type",
-    "load_resume_settings",
-    "resolve_config_path",
-    "launch_all_stages",
-    "launch_sweep",
-    "plot_sweep_results",
-    "run_trial",
-    "save_search_space",
-    "to_ray_tune",
-    "train_trial",
-    "write_results_csv",
     "compute_quality_scores",
     "create_ray_tuner",
     "detect_gpu_info",
@@ -123,6 +96,16 @@ __all__ = [
     "discover_and_rank_trials",
     "evaluate_trials_parallel",
     "export_best_trial",
+    "launch_all_stages",
+    "launch_sweep",
+    "load_resume_settings",
     "load_scoring_config",
+    "plot_sweep_results",
+    "resolve_config_path",
     "run_ray_sweep",
+    "run_trial",
+    "save_search_space",
+    "to_ray_tune",
+    "train_trial",
+    "write_results_csv",
 ]
